+++ src/cli/commands/repo/view.rs
@@ -17,10 +17,14 @@
 use std::fmt;
 
 use clap::Args;
-use nostr::nips::nip19::Nip19Coordinate;
 
 use crate::{
-    cli::{CliOptions, CommandRunner, parsers},
+    cli::{
+        CliConfig,
+        CliOptions,
+        CommandRunner,
+        types::{NaddrOrSet, OptionNaddrOrSetVecExt, RelayOrSetVecExt},
+    },
     error::N34Result,
     nostr_utils::{NostrClient, traits::NaddrsUtils, utils},
 };
@@ -28,20 +32,25 @@ use crate::{
 /// Arguments for the `repo view` command
 #[derive(Args, Debug)]
 pub struct ViewArgs {
-    /// Repository address in `naddr` format or `<nip5>/repo_id`. e.g.
-    /// `4rs.nl/n34` and `_@4rs.nl/n34`
+    /// Repository address in `naddr` format (`naddr1...`), NIP-05 format
+    /// (`4rs.nl/n34` or `_@4rs.nl/n34`), or a set name like `kernel`.
     ///
-    /// If not provided, `n34` will look for the `nostr-address` file.
-    #[arg(value_name = "NADDR-OR-NIP05", long = "repo", value_parser = parsers::repo_naddr)]
-    naddrs: Option<Vec<Nip19Coordinate>>,
+    /// If omitted, looks for a `nostr-address` file.
+    #[arg(value_name = "NADDR-NIP05-OR-SET", long = "repo")]
+    naddrs: Option<Vec<NaddrOrSet>>,
 }
 
 impl CommandRunner for ViewArgs {
     async fn run(self, options: CliOptions) -> N34Result<()> {
         // FIXME: The signer is not required here
 
-        let naddrs = utils::naddrs_or_file(self.naddrs, &utils::nostr_address_path()?)?;
-        let client = NostrClient::init(&options).await;
+        let config = CliConfig::load_toml(&options.config_path)?;
+        let naddrs = utils::naddrs_or_file(
+            self.naddrs.flat_naddrs(&config.sets)?,
+            &utils::nostr_address_path()?,
+        )?;
+        let relays = options.relays.clone().flat_relays(&config.sets)?;
+        let client = NostrClient::init(&options, &relays).await;
         client.add_relays(&naddrs.extract_relays()).await;
 
         let repos = client.fetch_repos(&naddrs.into_coordinates()).await?;